backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
aiohttp==3.10.0
aiosqlite==0.20.0
openai==1.50.0
pytrends==4.9.2
praw==7.7.1
beautifulsoup4==4.12.3
httpx==0.27.0
apscheduler==3.10.4
pydantic==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==1.1.0
pdfplumber==0.11.0
python-multipart==0.0.9
apify-client==1.8.1
